Move `RustdocTheme` to `rustdoc_tests`
